refactor: update imports for type definitions
